BUGFIX: Fixed a bug in locale fallback on toArray()
The bug occurs when you call toArray() on a model where a translation
does not exist for the current app locale and you have a translation
for the fallback locale set in the translatable.config

In this case you expect toArray() to merge the $translatedProperties
with the fallback translation, but this does not happen.

Added test to confirm bug is solved.

###### Sample translatable config setup
``` php

return [
    'locales' => [
        'nb' => [
            ,
        ],
        'da' => [
            'DK',
        ]
    ],
    'locale_separator' => '_',
    'locale' => null,
    'use_fallback' => true,
    'fallback_locale' => 'nb',
    'translation_suffix' => 'Translation',
    'locale_key' => 'locale',
    'always_fillable' => false,

];

```

###### Matching php sample
``` php

// lets say we have a food model with a $translatedAttributes =
['name'];
Food::create([
    'id' => 1,
    'nb' => [
        'name' => 'test'
    ]
]);

app()->setLocale('da_DK');

$food = Food::find(1)->toArray();

// Because our config has locale_fallback set to 'nb' we expect $food
to be:

[
  'id' => 1,
  'name' => 'test'
]

// However because of the bug $food is just:

[
  'id' => 1,
]

```