Fix issue when trying to fetch a translation with a country based (e.g. 'pt-PT') and the only translation available is for the fallback locale.
Cover the scenario where the configured fallback locale is 'en', you are trying to fetch a translation for 'pt-PT'
but the model only have a translation to 'en'. As it was before the code would attempt to fetch a translation for 'pt' twice instead of trying 'pt' first and then the configured fallback locale.